+++ src/common/modules/browser/url.ts
@@ -2,6 +2,10 @@ import { isDevelopment } from '@common/modules/isDevelopment'
 import { targetBrowser } from '@common/modules/browser/targetBrowser'
 
 export function getImageURL(uri: string): string {
+    return getPublicURL(`images/${uri}`)
+}
+
+export function getPublicURL(uri: string): string {
     if (isDevelopment()) {
         return `${uri}`
     }