+++ src/popup/components/Views/GroupView/IsUnlockedBox.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import type { Group } from '@/types'
+import { ref } from 'vue'
 import { trans } from '@common/modules/trans'
 import { useGroupStore } from '@/stores/group'
 import { usePopupStore } from '@/stores/popup'
@@ -15,17 +16,31 @@ type Props = {
 
 const { group } = defineProps<Props>()
 const groupStore = useGroupStore()
-const { openPopup } = usePopupStore()
+const { openPopup, onClose } = usePopupStore()
+
+const newPassword = ref<boolean>(false)
 
 async function promptEnterPassword(): Promise<void> {
+    if (newPassword.value) {
+        await deletePasswordFromStorage(group.id)
+    }
+
     const pass = await getPasswordFromStorage(group.id)
 
-    if (!pass) {
-        showToast(trans('cant_remember_pass'), 'error', 4000)
-        openPopup('newPassword')
+    if (pass) {
+        await lockGroup(pass)
         return
     }
 
+    if (!newPassword.value) {
+        showToast(trans('cant_remember_pass'), 'error', 4000)
+    }
+
+    openPopup('newPassword')
+    onClose(async (newPass: string) => await lockGroup(newPass))
+}
+
+async function lockGroup(pass: string): Promise<void> {
     await groupStore.encryptGroupById(group.id, pass)
     await deletePasswordFromStorage(group.id)
 
@@ -39,20 +54,25 @@ async function promptEnterPassword(): Promise<void> {
     >
         <ShieldExclamationIcon class="w-8 h-8 text-red-400" />
 
-        <span class="text-sm">
-            {{ trans('private_group_unlocked') }}
-        </span>
+        <span class="text-sm">{{ trans('private_group_unlocked') }}</span>
 
-        <button
-            @click="promptEnterPassword"
-            :class="[
-                'bg-unsafe px-3 py-2 rounded-md',
-                'text-sm hover:bg-unsafe-hover text-bg font-semibold',
-                'flex items-center gap-2',
-            ]"
-        >
-            <LockClosedIcon width="18" height="18" />
-            {{ trans('lock') }}
-        </button>
+        <div class="w-52 flex flex-col items-center gap-1.5">
+            <button
+                @click="promptEnterPassword"
+                :class="[
+                    'bg-unsafe px-3 py-2 rounded-md w-32',
+                    'text-sm hover:bg-unsafe-hover text-bg font-semibold',
+                    'flex items-center gap-2 justify-center',
+                ]"
+            >
+                <LockClosedIcon width="18" height="18" />
+                {{ trans('lock') }}
+            </button>
+
+            <label class="flex gap-1.5 align-center opacity-80">
+                <input type="checkbox" v-model="newPassword" />
+                <small>{{ trans('new_password') }}</small>
+            </label>
+        </div>
     </div>
 </template>